line-ai-bot/requirements.txt
functions-framework
flask
requests
google-cloud-secret-manager
google-generativeai
gunicorn
line-bot-sdk
